+++ service/dkg_generate_deals.go
@@ -3,11 +3,98 @@ package service
 import (
 	"bytes"
 	"context"
+	"encoding/hex"
 
 	"github.com/pkg/errors"
+	log "github.com/sirupsen/logrus"
 
+	"github.com/piplabs/story-kernel/enclave"
 	pb "github.com/piplabs/story-kernel/types/pb/v0"
+
+	dkg "go.dedis.ch/kyber/v4/share/dkg/pedersen"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
+
+func (s *DKGServer) GenerateDeals(_ context.Context, req *pb.GenerateDealsRequest) (*pb.GenerateDealsResponse, error) {
+	codeCommitmentHex := hex.EncodeToString(req.GetCodeCommitment())
+
+	// Validate request
+	if err := validateGenerateDealsRequest(req); err != nil {
+		log.WithFields(log.Fields{
+			"round":           req.GetRound(),
+			"code_commitment": codeCommitmentHex,
+		}).Errorf("invalid request: %v", err)
+
+		return nil, status.Errorf(codes.InvalidArgument, "invalid request")
+	}
+
+	// Validate code commitment
+	if err := enclave.ValidateCodeCommitment(req.GetCodeCommitment()); err != nil {
+		log.Errorf("failed to validate code commitment: %v", err)
+
+		return nil, status.Errorf(codes.InvalidArgument, "failed to validate code commitment")
+	}
+
+	rc, err := s.GetOrLoadRoundContext(codeCommitmentHex, req.GetRound())
+	if err != nil {
+		log.WithFields(log.Fields{
+			"round":           req.GetRound(),
+			"code_commitment": codeCommitmentHex,
+		}).Errorf("failed to get or load roundContext: %v", err)
+
+		return nil, status.Errorf(codes.Internal, "failed to get or load roundContext")
+	}
+
+	if err := s.CachePID(codeCommitmentHex, req.Round, rc.Registrations); err != nil {
+		log.WithFields(log.Fields{
+			"round":           req.GetRound(),
+			"code_commitment": codeCommitmentHex,
+		}).Errorf("failed to cache PID: %v", err)
+
+		return nil, status.Errorf(codes.Internal, "failed to cache PID")
+	}
+
+	// Load DKG state from cache or rebuild from state
+	var distKeyGen *dkg.DistKeyGenerator
+	if !req.GetIsResharing() {
+		distKeyGen, err = s.GetInitDKG(codeCommitmentHex, req.GetRound(), rc.Network.GetThreshold(), rc.SortedPubKeys)
+		if err != nil {
+			log.Errorf("failed to setup initial round DKG: %v", err)
+
+			return nil, status.Errorf(codes.Internal, "failed to load or rebuild initial round DKG")
+		}
+	} else {
+		latest, err := s.QueryClient.GetLatestActiveDKGNetwork(context.Background())
+		if err != nil {
+			log.Errorf("failed to get latest active DKG network: %v", err)
+
+			return nil, status.Errorf(codes.Internal, "failed to get latest active DKG network")
+		}
+
+		distKeyGen, err = s.GetResharingPrevDKG(codeCommitmentHex, req.GetRound(), rc.Network.GetThreshold(), rc.SortedPubKeys, latest)
+		if err != nil {
+			log.Errorf("failed to setup prev round DKG: %v", err)
+
+			return nil, status.Errorf(codes.Internal, "failed to load or rebuild prev round DKG")
+		}
+	}
+
+	// Generate deals
+	deals, err := distKeyGen.Deals()
+	if err != nil {
+		log.Errorf("failed to generate encrypted deals: %v", err)
+
+		return nil, status.Errorf(codes.Internal, "failed to generate encrypted deals")
+	}
+
+	log.Info("Succeed to generate deals", "code_commitment", codeCommitmentHex, "round", req.GetRound())
+
+	// Set deals into response
+	resp := createGenerateDealsResponse(req.GetRound(), req.GetCodeCommitment(), deals)
+
+	return resp, nil
+}
 
 func validateGenerateDealsRequest(req *pb.GenerateDealsRequest) error {
 	if req.GetRound() == 0 {
